terraform/CloudFunctionsDemo/src2/requirements.txt
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
google-api-core==1.31.1
google-auth==1.34.0
google-auth-httplib2==0.1.0
google-cloud-bigquery==2.23.3
google-cloud-core==1.7.2
google-cloud-pubsub==2.7.0
google-cloud-storage==1.41.1
google-crc32c==1.1.2
google-resumable-media==1.3.3
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.39.0
httplib2==0.19.1
idna==3.2
libcst==0.3.20
mypy-extensions==0.4.3
packaging==21.0
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
typing-extensions==3.10.0.0
typing-inspect==0.7.1
uritemplate==3.0.1
urllib3==1.26.6
